backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# Data Sources
requests==2.31.0
aiohttp==3.9.1
yfinance==0.2.35
fredapi==0.5.1
alpha-vantage==2.3.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Scheduling & Background Jobs
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# AI Providers
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0
